// frontend/src/app/api/model/timetableHearingStatementDataProtection.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimetableHearingStatementDocument } from './timetableHearingStatementDocument';


export interface TimetableHearingStatementDataProtection { 
    /**
     * Object creation date
     */
    readonly creationDate?: string;
    /**
     * User creator
     */
    readonly creator?: string;
    /**
     * Last edition date
     */
    readonly editionDate?: string;
    /**
     * User editor
     */
    readonly editor?: string;
    /**
     * Technical identifier
     */
    readonly id?: number;
    /**
     * Statement does not contain personal data
     */
    statementAnonymous?: boolean;
    /**
     * Statement anonymized by canton
     */
    anonymousStatement?: string;
    /**
     * List of uploaded documents
     */
    documents?: Array<TimetableHearingStatementDocument>;
}